Make TabletMap return empty replica set when stale replica is found
Makes TabletMap return empty collection for some calls to `getReplicas()`.
If the current replica list for a tablet turns out to contain a host
that is not present in current Metadata then empty collection is returned
in an effort to misroute the query. This query will cause the tablet
information to be updated when the result with the up to date information
is received.
It is possible that the query will be randomly routed correctly, but this
case is significantly less likely with each subsequent query, although
this can depend on underlying load balancing policy.

This covers the node removal/replacement case of #378.