Allow additional `scylla.version` formats (#396)
Modifies version parsing logic to try passing the `scylla.version` to the
CCM when it can't be parsed as VersionNumber. If CCM manages to run `create`
with it, then the usual version number is fetched from
`ccm node versionfrombuild` output.

To do that we introduce a static version of `execute`. The method body is
nearly the same. Functionally should be the same. Previous version now calls
the static one.